Use the right method name for babi.table's cron and make m2o fields required.
We did not change the name of the method in cron's selection in a previous
commit. That did not cause issues in general because those cron jobs are
executed by overriding run_once() but the problem arised if for some reason the
ir.cron record was stored with the m2o field empty.

I that case super's method would call the compute() method, but it should have
been _compute(). This caused the whole cron system to fail and reexecute all
tasks due to limitations in the cron implementation as of version 7.2 and below.

With this commit we make m2o fields required so the situation cannot happen but
even then we still use the right method name for consistency.

We also use the queue name as set in the configuration.