Add support for parametrized tables.
Now a table can be created with filter with parameters. When computed, a wizard
is shown asking the user to introduce the required parameters and a new table
is then created and computed.

That table appears in the Parametrized tables tab and those are removed by a
new `clean` cron job.